Rename MLP to LogisticRegression
