+++ models/lr/model.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 
 
-class MLP(nn.Module):
+class LogisticRegression(nn.Module):
 
     def __init__(self, config):
         super().__init__()
